--- a/ledger/common/proof/proof.go
+++ b/ledger/common/proof/proof.go
@@ -17,9 +17,9 @@ func VerifyTrieProof(p *ledger.TrieProof, expectedState ledger.State) bool {
 		return false
 	}
 	// We start with the leaf and hash our way upwards towards the root
-	proofIndex := len(p.Interims) - 1                                                      // the index of the last non-default value furthest down the tree (-1 if there is none)
-	computed := ledger.ComputeCompactValue(hash.Hash(p.Path), p.Payload.Value, leafHeight) // we first compute the hash of the compact leaf (at height leafHeight)
-	for h := leafHeight + 1; h <= treeHeight; h++ {                                        // then, we hash our way upwards until we hit the root (at height `treeHeight`)
+	proofIndex := len(p.Interims) - 1                                                        // the index of the last non-default value furthest down the tree (-1 if there is none)
+	computed := ledger.ComputeCompactValue(hash.Hash(p.Path), p.Payload.Value(), leafHeight) // we first compute the hash of the compact leaf (at height leafHeight)
+	for h := leafHeight + 1; h <= treeHeight; h++ {                                          // then, we hash our way upwards until we hit the root (at height `treeHeight`)
 		// we are currently at a node n (initially the leaf). In this iteration, we want to compute the
 		// parent's hash. Here, h is the height of the parent, whose hash want to compute.
 		// The parent has two children: child n, whose hash we have already computed (aka `computed`);