--- a/ledger/trie.go
+++ b/ledger/trie.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 
 	"github.com/fxamacker/cbor/v2"
+
 	cryptoHash "github.com/onflow/flow-go/crypto/hash"
 	"github.com/onflow/flow-go/ledger/common/bitutils"
 	"github.com/onflow/flow-go/ledger/common/hash"