This PR replaces decoded payload key with encoded key buffer because decoded payload key is only used for migration and reports.  This change made ledger.Payload immutable.

Closes #2569
Updates #1744

### Goals

- (main goal) reduce memory required by EN (operational RAM).
- reduce number of allocations on the heap.
- have zero negative performance tradeoffs.

### Impact

- operational RAM should be reduced by dozens of GB (very roughly 25-50GB initially and more as data grows)
- eliminate over 1+ billion heap allocations for mtrie in memory (about 1+ billion allocs when mtrie is created and additional savings during activities that update mtrie)
- as positive side-effects, speedup
  - ledger update (see TrieUpdate benchstats)
  - EN startup
  - checkpoint (de)serialization
  - TrieProof (de)serialization, and etc.

### Example positive side-effect beyond operational RAM reduction

Benchstats are only for ledger update.  Other improvements are not benchmarked yet.

```
name          old time/op    new time/op    delta
TrieUpdate-4     439ms ± 2%     409ms ± 1%   -6.94%  (p=0.000 n=18+20)

name          old alloc/op   new alloc/op   delta
TrieUpdate-4    73.5MB ± 0%    34.1MB ± 0%  -53.60%  (p=0.000 n=20+20)

name          old allocs/op  new allocs/op  delta
TrieUpdate-4      187k ± 0%      147k ± 0%  -21.44%  (p=0.000 n=20+20)
```

### Caveats

- Updating benchmark results or adding missing benchmarks can be done at a later date.
- Custom functions for CBOR and JSON serialization was added because immutable fields are no longer exported, but the memory reduction and speedups in other ops outweigh this.

### Reviewers

This PR is fairly simple because the most important changes are in the small commit e35bd9449eae87af76ecbdb6422c2b5c5ed44ba9.

The large number of lines changed by other commit is to:
- commit b2517548083f005fb8321d6ba820c7ac885c957b - make ledger.Payload immutable
- commit 74ee21ddaacb4cdc200b7441ee9a6f49d53d0b9e - eliminate circular dependencies (moving ledger/common/encoding/encodiing.go to ledger/trie_encoder.go, split common/utils/testutils.go into 2 packages)